proxy_server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
google-auth>=2.0.0 # Or a specific version if preferred
requests>=2.0.0    # Required by google-auth for default transport